.vscode/my_project/requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysqlclient
python-dotenv
streamlit
